requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database (PostgreSQL only)
psycopg2-binary==2.9.9

# LLM (optional, for better intent extraction)
openai==1.3.0
langchain==0.0.340  # Optional, can remove if you want minimal

# Visualization
plotly==5.18.0
pandas==2.1.3

# Testing (optional)
pytest==7.4.3

# Total: ~10 packages, minimal dependencies